refactor: update type convert function
- remove unused functions
- remove unnecessary comments

+++ internal/model/admin/account_transaction/get_all.go
@@ -21,8 +21,8 @@ type GetAllItem struct {
 	ID              string `json:"id"`
 	TransactionDate string `json:"transactionDate"`
 	Type            string `json:"type"`
-	Amount          int    `json:"amount"`
-	Balance         int    `json:"balance"`
+	Amount          int64  `json:"amount"`
+	Balance         int64  `json:"balance"`
 	Note            string `json:"note"`
 	CreatedAt       string `json:"createdAt"`
 	UpdatedAt       string `json:"updatedAt"`
